% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_data.R
\name{Data}
\alias{Data}
\title{Generate Time to event data using the framework of Relative time method}
\usage{
Data(a, f, beta0, median_c, p1, p2, RT_P1, RT_P2, scenario)
}
\arguments{
\item{a}{numeric variable}

\item{f}{numeric variable}

\item{beta0}{numeric variable}

\item{median_c}{numeric variable}

\item{p1}{numeric variable}

\item{p2}{numeric variable}

\item{RT_P1}{numeric variable}

\item{RT_P2}{numeric variable}

\item{scenario}{numeric variable between 1 and 4}
}
\value{

}
\description{
Generate Survival data for Sample size re-estimation method using the framework of Relative time method
}
\examples{
Data(a = 12,f = 6,beta0 = 0.25,median_c = 4,p1 = 0.10,p2 = 0.90,RT_P1 = 1.52,RT_P2 = 1.98,scenario = 1)
set.seed(222)
interim <- Data(a = 12,f = 6,beta0 = 0.25,median_c = 4,p1 = 0.10,p2 = 0.90,RT_P1 = 1.52,RT_P2 = 1.98,scenario = 1)
sum(interim$new.event)
\dontrun{
set.seed(222)
interim <- Data(a = 12,f = 6,beta0 = 0.25,median_c = 4,p1 = 0.10,p2 = 0.90,RT_P1 = 1.52,RT_P2 = 1.98,scenario = 1)
sum(interim$new.event)
#' }

}
