% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reestim_IPS.R
\name{event}
\alias{event}
\title{Calculate the event using the IPS approach}
\usage{
event(beta0, Interim.event, fixed_sample, scenario)
}
\arguments{
\item{beta0}{numeric variable}

\item{Interim.event}{numeric variable}

\item{fixed_sample}{numeric variable}

\item{scenario}{numeric variable between 1 and 4}
}
\value{

}
\description{
Here, we used the IPS approach to re-estimate the sample size based on the relative time method
}
\examples{
event(beta0 = 0.462,Interim.event = 53,fixed_sample = 92,scenario = 1)
}
