% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_R.R
\name{interim_data}
\alias{interim_data}
\title{Generate Interim survival data based on Known Weibull control shape parameter}
\usage{
interim_data(beta_c, beta_t, median_c, median_t)
}
\arguments{
\item{beta_c}{numeric variable}

\item{beta_t}{numeric variable}

\item{median_c}{numeric variable}

\item{median_t}{numeric variable}
}
\value{
Generate Interim survival data
}
\description{
Generate Interim survival data based on Known Weibull control shape parameter
}
\examples{
sample_final <- sample_size(alpha = 0.05,sides = 1,beta0  = 0.25,median_c = 4,p1 = 0.1,p2 = 0.9,RT_P1 = 1.52,RT_P2 = 1.98,qmin = 0.001,qmax = 0.999,Power = 0.8,r = 1,a = 12,f = 12,p_event=0.8)

sample_final

sample_int <- sample_size(alpha = 0.05,sides = 1,beta0  = 0.25,median_c = 4,p1 = 0.1,p2 = 0.9,RT_P1 = 1.52,RT_P2 = 1.98,qmin = 0.001,qmax = 0.999,Power = 0.8,r = 1,a = 12,f = 6,p_event=0.8)
sample_int

data_int <- interim_data(beta_c = 0.25,beta_t =0.2447551, median_c = 4,median_t = 6)

\dontrun{
sample_final <- sample_size(alpha = 0.05,sides = 1,beta0  = 0.25,median_c = 4,p1 = 0.1,p2 = 0.9,RT_P1 = 1.52,RT_P2 = 1.98,qmin = 0.001,qmax = 0.999,Power = 0.8,r = 1,a = 12,f = 12,p_event=0.8)

sample_final

sample_int <- sample_size(alpha = 0.05,sides = 1,beta0  = 0.25,median_c = 4,p1 = 0.1,p2 = 0.9,RT_P1 = 1.52,RT_P2 = 1.98,qmin = 0.001,qmax = 0.999,Power = 0.8,r = 1,a = 12,f = 6,p_event=0.8)
sample_int

data_int <- interim_data(beta_c = 0.25,beta_t =0.2447551, median_c = 4,median_t = 6)

}

}
