% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{Interim_test}
\alias{Interim_test}
\title{Title Stochastic curtailment test calculation}
\usage{
Interim_test(alpha, beta_c, scenario)
}
\arguments{
\item{alpha}{numeric variable}

\item{beta_c}{numeric variable}

\item{scenario}{numeric variable}
}
\description{
Title Stochastic curtailment test calculation
}
\examples{
sample_final <- sample_size(alpha = 0.05,sides = 1,beta0  = 0.25,median_c = 4,p1 = 0.1,p2 = 0.9,RT_P1 = 1.52,RT_P2 = 1.98,qmin = 0.001,qmax = 0.999,Power = 0.8,r = 1,a = 12,f = 12,p_event=0.8)

sample_final

sample_int <- sample_size(alpha = 0.05,sides = 1,beta0  = 0.25,median_c = 4,p1 = 0.1,p2 = 0.9,RT_P1 = 1.52,RT_P2 = 1.98,qmin = 0.001,qmax = 0.999,Power = 0.8,r = 1,a = 12,f = 6,p_event=0.8)
sample_int

data_int <- interim_data(beta_c = 0.25,beta_t =0.2447551, median_c = 4,median_t = 6)

result = Interim_test(alpha = 0.05*0.90,beta_c = 0.25,scenario =1)
result

\dontrun{
sample_final <- sample_size(alpha = 0.05,sides = 1,beta0  = 0.25,median_c = 4,p1 = 0.1,p2 = 0.9,RT_P1 = 1.52,RT_P2 = 1.98,qmin = 0.001,qmax = 0.999,Power = 0.8,r = 1,a = 12,f = 12,p_event=0.8)

sample_final

sample_int <- sample_size(alpha = 0.05,sides = 1,beta0  = 0.25,median_c = 4,p1 = 0.1,p2 = 0.9,RT_P1 = 1.52,RT_P2 = 1.98,qmin = 0.001,qmax = 0.999,Power = 0.8,r = 1,a = 12,f = 6,p_event=0.8)
sample_int

data_int <- interim_data(beta_c = 0.25,beta_t =0.245, median_c = 4,median_t = 6) # Secnerio 2: They are different : RT_P1 = 1.52 AND RT_P2 = 1.98

result = Interim_test(alpha = 0.05*0.90,beta_c = 0.25,scenario =1)
result
}



sample_int
}
