% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Relative_Size.R
\name{Relative_size}
\alias{Relative_size}
\title{Calculate Sample size using the Relative Time method under Non-Proportional Hazard and Non-Proportional Time scenarios}
\usage{
Relative_size(
  alpha,
  sides,
  beta0,
  median_c,
  p1,
  p2,
  p_user = NULL,
  RT_P1,
  RT_P2,
  qmin,
  qmax,
  Power,
  r,
  a,
  f,
  p_event
)
}
\arguments{
\item{alpha}{numeric variable}

\item{sides}{binary variable}

\item{beta0}{numeric variable}

\item{median_c}{numeric variable}

\item{p1}{numeric variable}

\item{p2}{numeric variable}

\item{p_user}{numeric variable}

\item{RT_P1}{numeric variable}

\item{RT_P2}{numeric variable}

\item{qmin}{numeric variable}

\item{qmax}{numeric variable}

\item{Power}{numeric variable}

\item{r}{numeric variable}

\item{a}{numeric variable}

\item{f}{numeric variable}

\item{p_event}{numeric variable}
}
\value{
Sample size Calculation using the Relative Time method proposed by Phadnis and Mayo (2021).
}
\description{
Calculate Sample size using the Relative Time method under Non-Proportional Hazard and Non-Proportional Time scenarios
}
\examples{
Relative_size(alpha = 0.05,sides = 1,beta0  = 1,median_c = 4,p1 = 0.1,p2 = 0.9,RT_P1 = 1.52,RT_P2 = 1.98,qmin = 0.001,qmax = 0.999,Power = 0.8,r = 1,a = 12,f = 12,p_event=0.8)
\dontrun{
Relative_size(alpha = 0.05,sides = 1,beta0  = 1,median_c = 4,p1 = 0.1,p2 = 0.9,RT_P1 = 1.52,RT_P2 = 1.98,qmin = 0.001,qmax = 0.999,Power = 0.8,r = 1,a = 12,f = 12,p_event=0.8)
}

}
