% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reestim_pred.R
\name{search.best.n.pred}
\alias{search.best.n.pred}
\title{Predicitve Power approach to sample size recalculation}
\usage{
search.best.n.pred(nstart, nmax, D, beta0, scenario)
}
\arguments{
\item{nstart}{Numeric values}

\item{nmax}{Numeric values}

\item{D}{Numeric values}

\item{beta0}{Numeric values}

\item{scenario}{Numeric values between 0 and 4}
}
\value{

}
\description{
Predicitve Power approach to sample size recalculation
}
\examples{
search.best.n.pred(nstart = 60,nmax = 300,D = 82,beta0 = 0.459,scenario = 1)

}
