% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type_1_error.R
\name{type.1.err}
\alias{type.1.err}
\title{Operating characteristics of the Relative time method}
\usage{
type.1.err(num_sims, a, f, Hyp, p1, p2, beta0, scenario)
}
\arguments{
\item{num_sims}{numeric variable}

\item{a}{numeric variable}

\item{f}{numeric variable}

\item{Hyp}{Binary variable (0/1); 0 indicate the Type I error rate and 1 Indicate the associated Power}

\item{p1}{numeric variable}

\item{p2}{numeric variable}

\item{beta0}{numeric variable}

\item{scenario}{numeric variable between 1 and 4}
}
\value{

}
\description{
This function calculate the Type I error rate based on this method and associated Power under various scenarios.
}
\examples{
num_sims <- 10000
sims <- plyr::ldply(1:num_sims, type.1.err, a = 12, f=12, Hyp=0,p1=0.1, p2=0.9,beta0 = 0.25,scenario = 1)
result <- apply(sims, MARGIN = 2 , FUN=mean, na.rm = TRUE)
result
\dontrun{
num_sims <- 10000
sims <- plyr::ldply(1:num_sims, type.1.err, a = 12, f=12, Hyp=0,p1=0.1, p2=0.9,beta0 = 0.25,scenario = 1)
result <- apply(sims, MARGIN = 2 , FUN=mean, na.rm = TRUE)
result
}

}
